refactor: add ShieldManager class to separate shield logic in Boss and ShieldImage
- created `ShieldManager` class to handle shield's state, activation, deactivation, and visibility
- `ShieldManager` improves separation of concerns and simplify the Boss and ShieldImage class
