+++ src/main/java/com/example/demo/actors/planes/Boss.java
@@ -3,6 +3,7 @@ package com.example.demo.actors.planes;
 import com.example.demo.actors.core.ActiveActorDestructible;
 import com.example.demo.actors.projectiles.BossProjectile;
 import com.example.demo.actors.shield.ShieldImage;
+import com.example.demo.actors.shield.ShieldManager;
 
 import java.util.*;
 
@@ -13,7 +14,7 @@ public class Boss extends FighterPlane {
 	private static final double INITIAL_Y_POSITION = 400;
 	private static final double PROJECTILE_Y_POSITION_OFFSET = 75.0;
 	private static final double BOSS_FIRE_RATE = .04;
-	private static final double BOSS_SHIELD_PROBABILITY = .002;
+	//private static final double BOSS_SHIELD_PROBABILITY = 0.02;
 	private static final int IMAGE_HEIGHT = 100;
 	private static final int VERTICAL_VELOCITY = 8;
 	private static final int HEALTH = 1;
@@ -22,23 +23,26 @@ public class Boss extends FighterPlane {
 	private static final int MAX_FRAMES_WITH_SAME_MOVE = 10;
 	private static final int Y_POSITION_UPPER_BOUND = -30;
 	private static final int Y_POSITION_LOWER_BOUND = 475;
-	private static final int MAX_FRAMES_WITH_SHIELD = 500; // active for ~8.33 seconds
+	//private static final int MAX_FRAMES_WITH_SHIELD = 500; // active for ~8.33 seconds
 	private final List<Integer> movePattern;
 	private boolean isShielded;
 	private int consecutiveMovesInSameDirection;
 	private int indexOfCurrentMove;
-	private int framesWithShieldActivated;
+	//private int framesWithShieldActivated;
 	private final ShieldImage shieldImage;
+	private final ShieldManager shieldManager;
+	//private final ShieldFactory shieldFactory;
 
 	public Boss() {
 		super(IMAGE_NAME, IMAGE_HEIGHT, INITIAL_X_POSITION, INITIAL_Y_POSITION, HEALTH);
-		movePattern = new ArrayList<>();
+        movePattern = new ArrayList<>();
 		consecutiveMovesInSameDirection = 0;
 		indexOfCurrentMove = 0;
-		framesWithShieldActivated = 0;
+		//framesWithShieldActivated = 0;
 		isShielded = false;
 		initializeMovePattern();
 		shieldImage = new ShieldImage(INITIAL_X_POSITION, INITIAL_Y_POSITION);
+		shieldManager = new ShieldManager(shieldImage, this);
 	}
 
 	//applies next movement to Boss's position
@@ -51,14 +55,14 @@ public class Boss extends FighterPlane {
 		if (currentPosition < Y_POSITION_UPPER_BOUND || currentPosition > Y_POSITION_LOWER_BOUND) {
 			setTranslateY(initialTranslateY);
 		}
-		updateShieldPosition();
+		shieldManager.updateShieldPosition();
 		updateHitboxPosition();
 	}
 	
 	@Override
 	public void updateActor() {
 		updatePosition();
-		updateShield();
+		shieldManager.updateShield();
 	}
 
 	@Override
@@ -69,7 +73,7 @@ public class Boss extends FighterPlane {
 	@Override
 	public void takeDamage() {
 		//only takes damage if not shielded
-		if (!isShielded) {
+		if (!shieldManager.isShielded) {
 			super.takeDamage();
 		}
 	}
@@ -83,29 +87,6 @@ public class Boss extends FighterPlane {
 			movePattern.add(ZERO); //stationary
 		}
 		Collections.shuffle(movePattern);
-	}
-
-    public void updateShieldPosition(){
-        double shieldXOffset = -getImageWidth() * 0.1;
-        shieldImage.setLayoutX(getLayoutX() + shieldXOffset);
-        shieldImage.setLayoutY(getLayoutY() + getTranslateY() - getImageHeight() / 4.0);
-    }
-
-	private void updateShield() {
-		if (isShielded) {
-			framesWithShieldActivated++;
-			System.out.println("Shield Active: " + isShielded + ", Frames: " + framesWithShieldActivated);
-		}
-		else if (shieldShouldBeActivated()) {
-			activateShield();
-			shieldImage.showShield(); // show shield when activated
-			System.out.println("Shield Active: " + isShielded + ", Frames: " + framesWithShieldActivated);
-		}
-		if (shieldExhausted()) {
-			deactivateShield();
-			shieldImage.hideShield(); // hide shield when exhausted
-			System.out.println("Shield Active: " + isShielded + ", Frames: " + framesWithShieldActivated);
-		}
 	}
 
 	//determines next vertical movement for Boss based on movePattern
@@ -127,32 +108,16 @@ public class Boss extends FighterPlane {
 		return Math.random() < BOSS_FIRE_RATE;
 	}
 
-	private boolean shieldShouldBeActivated() {
-		return Math.random() < BOSS_SHIELD_PROBABILITY;
-	}
-
-	private boolean shieldExhausted() {
-		return framesWithShieldActivated == MAX_FRAMES_WITH_SHIELD;
-	}
-
-	private void activateShield() {
-		isShielded = true;
-	}
-
-	private void deactivateShield() {
-		isShielded = false;
-		framesWithShieldActivated = 0;
-	}
-
-	public ShieldImage getShieldImage(){
-		return shieldImage;
+	@Override
+	public double getImageHeight(){
+		return getImage().getHeight();
 	}
 
 	public double getImageWidth(){
 		return getImage().getWidth();
 	}
 
-	public double getImageHeight(){
-		return getImage().getHeight();
+	public ShieldImage getShieldImage(){
+		return shieldImage;
 	}
 }
